Add the game of Rock-Paper-Scissors-Lizard-Spock. (#15)
This additional game acts as a second implementation to see the common patterns to implement, as well as features to add.

As a result, we must try to implement an helper chan to order and sequence inputs from users.
We also need to add arguments when creating a gathering. For RPSLS, it is the target score of the game.