Add options to gathering. (#16)
This allows specifying the number of players, target scores for games, ...

This requires a change of the syntax when creating games. Instead of a simple map, the factory protocol was extended to the whole game.

Note 1: this finally commits the tests for RPSLS.
Note 2: this reviewed the system to run tests, to make it more REPL-friendly.